title page and about page finished, but not mobile-friendly yet
