--- src/Components/Footer/index.css
@@ -1,67 +1,0 @@
-footer {
-    height: 105px;
-    background-color: rgb(46,53,70);
-    color: white;
-    position: relative;
-    padding-top: 1px;
-    text-align: center;
-    font-size: 30px;
-}
-
-footer > div {
-    margin-top: 10px;
-}
-
-footer .border1 {
-    grid-area: s1;
-}
-
-footer .border2 {
-    grid-area: s2;
-}
-
-footer .border1,
-footer .border2 {
-    border-top: 1px solid white;
-    margin-top: 12.8px;
-}
-
-footer .footerGrid {
-    display: grid;
-    grid-template-columns: 1fr 130px 1fr;
-    grid-template-areas:
-        "s1 m s2";
-}
-
-footer h3 {
-    text-align: center;
-    padding-top: 0px;
-    margin-bottom: 15px;
-    margin-top: 0;
-    font-weight: 500;
-    letter-spacing: 0.7px;
-    font-size: 22px;
-    padding-right: 17px;
-    padding-left: 17px;
-    width: 130px;
-    grid-area: m;
-}
-
-footer a {
-    margin-bottom: 0px;
-    color: white;
-    margin-left: 7px;
-    margin-right: 7px;
-}
-
-footer a:hover {
-    -webkit-transition-duration: 300ms;
-            transition-duration: 300ms;
-    color: rgb(208,157,85);
-}
-
-footer i {
-    text-align: center;
-    font-size: 30px;
-    padding: 3px;
-}
