+++ src/Components/Navbar/index.css
@@ -1,63 +1,55 @@
 nav {
-    height: 40px;
-    text-shadow: 1px 1px black;
+    z-index: 2;
+    position: fixed;
+	right: 20px;
+	top: 20px;
+    height: 60px;
+	margin: 0px;
+	font-size: 28px;
+    text-align: center;
+    padding: 16px 32px;
+    padding-top: 14px;
+    border-radius: 8px;
+    background-color: var(--navbar);
+    opacity: 0.7;
+}
+
+nav * {
+    color: white;
+}
+
+nav:hover,
+nav:focus {
+    cursor: pointer;
+    background-color: var(--navbarHover);
 }
 
 nav ul {
     list-style: none;
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-pack: space-evenly;
-        -ms-flex-pack: space-evenly;
-            justify-content: space-evenly;
-            align-items: center;
-    margin: 0 0;
-    padding: 0;
-    font-size: 17px;
-    height: 40px;
+	position: absolute;
+	background-color: transparent;
+	min-width: 150px;
+	text-align: center;
+    top: 32px;
+    right: 0px;
 }
 
-nav li {
-    z-index: 10;
-    width: 25%;
-    text-align: center;
-}
-
-nav a:hover {
-    -webkit-transition-duration: 300ms;
-            transition-duration: 300ms;
-    background-color: white;
-    color: rgb(40, 61, 112);
-    text-shadow: none;
-}
-
-nav a {
-    color: white;
+nav ul a {
     text-decoration: none;
-    padding: 9.5px 19px;
-    font-weight: 300;
-    letter-spacing: 0.6px;
-    font-size: 17px;
-    font-family: "Muli", serif;
-    line-height: initial;
-    text-align: center;
-    margin: auto;
 }
 
-nav .overlay {
-    background-color: rgb(46,53,70);
-    opacity: 0.75;
-    width: 100%;
-    height: 40px;
-    position: absolute;
-    top: 0px;
-    z-index: 9;
+nav ul li {
+	padding: 8px 16px;
+	box-shadow: 0 1px 1px 0 rgba(0, 0, 0, 0.3), 0 1px 3px 0 rgba(0, 0, 0, 0.33);
+    background-color: var(--navbar);
+    font-size: 18px;
+    border-radius: 8px;
+    margin-top: 10px;
+    margin-bottom: 10px;
+    font-weight: 400;
 }
 
-@media (max-width: 425px) {
-    nav a {
-        padding: 11px 10px;
-        font-size: 14px;
-    }
+nav ul li:hover {
+    background-color: var(--navbarHover);
+    cursor: pointer;
 }
