--- src/Pages/Home/Thumbnail/index.css
@@ -1,85 +1,0 @@
-/* Thumbnails */
-
- .threeThumbnails .thumbnail {
-     width: 320px;
-     padding: 30px 20px;
- }
-
- .threeThumbnails .thumbnail:hover {
-     background-color: rgb(208,157,85);
-     cursor: pointer;
-     -webkit-transition-duration: 500ms;
-             transition-duration: 500ms;
- }
-
- .threeThumbnails .thumbnail:hover div {
-     background-color: white;
- }
-
-
- .threeThumbnails .thumbnailImg {
-     margin: auto;
-     width: 120px;
-     height: 120px;
-     position: relative;
-     border-radius: 50%;
-     overflow: hidden;
- }
-
- .threeThumbnails .thumbnailImg img {
-     width: 105%;
-     height: 105%;
-     border-radius: 50%;
-     border: 1px solid black;
-     z-index: 2;
-     position: absolute;
-     left: 50%;
-     top: 50%;
-     -webkit-transform: translate(-50%, -50%);
-             transform: translate(-50%, -50%);
- }
-
- .threeThumbnails .thumbnail a h5 {
-     line-height: 29px;
-     color: rgb(46,53,70);
- }
-
- @media (max-width: 679px) {
-     .threeThumbnails .thumbnail {
-         width: 100%;
-     }
- }
-
- @media (max-width: 480px) {
-     .threeThumbnails .thumbnail h3 {
-         font-size: 32px;
-     }
-
-     .threeThumbnails .thumbnail h5 {
-         font-size: 20px;
-     }
- }
-
-/* Specific Changes */
-
- .threeThumbnails #Projects:hover *,
- .threeThumbnails #About:hover *,
- .threeThumbnails #Resume:hover *  {
-     -webkit-transition-duration: 500ms;
-             transition-duration: 500ms;
-     color: rgb(46,53,70);
- }
-
- .threeThumbnails #Projects img {
-     opacity: 0.8;
- }
-
- @media (max-width: 679px) {
-     .threeThumbnails #About {
-         background-color: rgb(224, 221, 213);
-     }
-
-     .threeThumbnails #About:hover {
-         background-color: rgb(208,157,85);
-     }
- }
