--- src/Pages/Projects/Project/index.css
@@ -1,140 +1,0 @@
-/* Project Entry Formatting */
-
-.projectEntry {
-    padding: 20px 15px;
-    margin: 0;
-}
-
-.projectEntry h2,
-.projectEntry h3 {
-    text-align: center;
-}
-
-.projectEntry h2 {
-    margin-top: 0px;
-    margin-bottom: 0px;
-    font-size: 30px;
-    color: rgb(112,53,41);
-    font-weight: 400;
-}
-
-.projectEntry h3 {
-    margin-top: 10px;
-    margin-bottom: 0px;
-    font-size: 24px;
-    font-family: "Montserrat", sans-serif;
-    font-weight: 300;
-    color: rgb(46,53,70);
-}
-
-.projectEntry h3 a,
-.projectEntry h2 a {
-    margin-left: 4px;
-    padding: 3px 4px;
-}
-
-.projectEntry h2 a {
-    padding: 5px 10px;
-    text-decoration: none;
-    color: rgb(112,53,41);
-    font-weight: 400;
-}
-
-.projectEntry h3 a {
-    color: rgb(46,53,70);
-    text-decoration: none;
-}
-
-.projectEntry h2 a:hover span,
-.projectEntry h3 a:hover span {
-    text-decoration: underline;
-}
-
-.projectEntry h5 {
-    padding-left: 20px;
-}
-
-.projectEntry h5,
-.projectEntry li {
-    font-size: 18px;
-    font-weight: 400;
-    font-family: "Montserrat", sans-serif;
-}
-
-.projectEntry h5 span {
-    font-weight: 600;
-    font-family: "Montserrat", sans-serif;
-    font-size: 18px;
-    color: rgb(40,61,112);
-}
-
-.projectEntry h6 {
-    text-align: center;
-    font-weight: 400;
-    font-family: "Montserrat", sans-serif;
-    font-size: 15px;
-    margin-top: 10px;
-    margin-bottom: 0;
-    display: none;
-    font-weight: 600;
-    color: rgb(180, 110, 20);
-}
-
-.projectEntry:hover h2,
-.projectEntry:hover h3,
-.projectEntry:hover h2 a,
-.projectEntry:hover h3 a {
-    color: rgb(180, 110, 20);
-    -webkit-transition-duration: 300ms;
-            transition-duration: 300ms;
-}
-
-.projectEntry:hover {
-    cursor: pointer;
-}
-
-.projectEntry:hover h6 {
-    display: block;
-}
-
-@media (max-width: 750px) {
-    .projectEntry h2 {
-        font-size: 22px;
-    }
-
-    .projectEntry h3 {
-        font-size: 18px;
-    }
-
-    .projectEntry h5,
-    .projectEntry li,
-    .projectEntry h5 span {
-        font-size: 16px;
-        line-height: 26px;
-    }
-}
-
-@media (max-width: 550px) {
-    .projectEntry ul {
-        padding-left: 20px;
-    }
-
-    .projectEntry h5 {
-        padding-left: 0px;
-    }
-}
-
-/* Image Formatting */
-
-.projectEntry .projectImages {
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -ms-flex-wrap: wrap;
-        flex-wrap: wrap;
-    -webkit-box-pack: justify;
-        -ms-flex-pack: justify;
-            justify-content: space-between;
-    margin-left: 10px;
-    margin-right: 10px;
-}
