+++ src/Sections/Intro/index.css
@@ -1,0 +1,78 @@
+/* Main Page */
+
+.intro {
+    height: 100vh;
+    width: 100%;
+    position: relative;
+}
+
+.intro .introImg {
+    background: -webkit-gradient(linear, left top, left bottom, from(rgba(0, 0, 0, 0.0)), to(rgba(0, 0, 0, 0.4))), url("../../Images/intro.jpg") no-repeat center center;
+    background: linear-gradient(rgba(0, 0, 0, 0.0), rgba(0, 0, 0, 0.4)), url("../../Images/intro.jpg") no-repeat center center;
+    z-index: 0;
+    background-size: cover;
+    height: 100%;
+    width: 100%;
+    position: absolute;
+    top: 0;
+    left: 0;
+}
+
+.intro .introText {
+    position: absolute;
+    z-index: 1;
+    text-align: center;
+    top: 50%;
+    height: 50%;
+    width: 100%;
+    position: absolute;
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    color: white;
+    font-family: 'Roboto Slab', serif;
+}
+
+.intro .introText h1,
+.intro .introText h2 {
+    margin: 0px;
+    font-weight: 200;
+}
+
+.intro h1 {
+    font-size: 130px;
+}
+
+.intro h2 {
+    font-size: 60px;
+}
+
+@media (max-width: 800px) {
+    .intro .introText h1 {
+        font-size: 80px;
+    }
+
+    .intro .introText h2 {
+        font-size: 40px;
+    }
+}
+
+@media (max-width: 650px) {
+    .intro .introText h1 {
+        font-size: 60px;
+    }
+
+    .intro .introText h2 {
+        font-size: 30px;
+    }
+}
+
+@media (max-width: 480px) {
+    .intro .introText h1 {
+        font-size: 45px;
+    }
+
+    .intro .introText h2 {
+        font-size: 24px;
+    }
+}
