+++ src/Sections/Intro/index.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import './index.css'
+
+const Intro = () => {
+    return (
+        <div className="intro">
+            <div className="introText">
+                <h1>Jared Ryan</h1>
+                <h2>Software and Web Developer</h2>
+            </div>
+            <div className="introImg"></div>
+        </div>
+    );
+}
+
+export default Intro;
