+++ src/index.css
@@ -1,21 +1,27 @@
+:root {
+    --navbar: black;
+    --navbarHover: rgb(112, 100, 98);
+}
+
 * {
     -webkit-box-sizing: border-box;
             box-sizing: border-box;
-    font-family: "Muli", serif;
+    /* font-family: "Muli", serif; */
 }
 
 body {
     margin: 0;
     padding: 0;
-    font-family: sans-serif;
-    background-color: rgb(46,53,70);
 }
 
 p, li {
     font-size: 18px;
-    /* font-weight: 300; */
-    font-family: "Montserrat", sans-serif;
+    font-family: "Muli", sans-serif;
     line-height: 28px;
+}
+
+h1, h2 {
+    font-family: 'Roboto Slab', serif;
 }
 
 @media (max-width: 750px) {
